+++ src/components/ui/LoadingTransition.tsx
@@ -2,9 +2,11 @@ import React from "react";
 
 const LoadingTransition: React.FC = () => {
   return (
-    <div className="fixed top-0 left-0 w-full h-full flex justify-center items-center bg-white z-50">
-      <div className="border-t-8 border-gray-300 border-solid border-b-8 border-blue-500 rounded-full w-16 h-16 animate-spin"></div>
-    </div>
+<div className="fixed top-0 left-0 w-full h-full flex justify-center items-center bg-white z-50">
+  <div className="w-16 h-16 border-8 border-blue-500 border-solid rounded-full animate-spin"></div>
+  <div className="w-16 h-16 border-8 border-blue-300 border-solid rounded-full absolute"></div>
+</div>
+
   );
 };
